pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "surface-defect-research"
version = "0.1.0"
description = "Research pipeline for surface defect recognition with synthetic data augmentation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "torch>=2.2",
  "torchvision>=0.17",
  "Pillow>=10.0",
  "matplotlib>=3.9",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
